pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "scikit-build>=0.13",
    "cmake>=3.18",
    "ninja",
]
build-backend = "madrona-py-build"
backend-path = ["external/madrona/py"]
wheel-directory = "build"

[project]
name = "gpudrive"
version = "0.3.0"
authors = [
    {name = "Saman Kazemkheni", email = "skazemkhani@gmail.com"},
    {name = "Aarav Pandya", email = "pandya.aarav.97@gmail.com"},
    {name = "Daphne Cornelisse", email = "cor.daphne@gmail.com"},
    {name = "Eugene Vinitsky", email ="vinitsky.eugene@gmail.com"},
]
readme = "README.md"
description = "A GPU-accelerated, multi-agent driving simulator"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.4",
    "gymnasium",
    "pygame",
    "matplotlib==3.9",
    "pandas",
    "python-box==7.2.0",
    "typer",
    "pyyaml", 
    "mediapy",
    "wandb",
    "torch>=2.2.1",
    "seaborn",
    "safetensors",
]

[project.optional-dependencies]
pufferlib = [
    "pufferlib>=2.0.6",
]
sb3 = [
    "stable-baselines3==2.3.2",
]

test = [
    "pytest>=8.2.1",
]

[tool.madrona.packages.madrona_gpudrive]
ext-only = true
ext-out-dir = "build"

[tool.madrona.packages.gpudrive]
path = "gpudrive"

[project.packages]
include = [
    "baselines",
    "baselines/*",
]